+++ modules/Sources/UIComponents/Tooltip/Tooltip.swift
@@ -39,12 +39,12 @@ public struct Tooltip: View {
                 VStack(alignment: .leading) {
                     Text(title)
                         .font(.custom(FontFamily.Inter.semiBold.name, size: 16))
-                        .zForegroundColor(Design.Text.primary)
+                        .zForegroundColor(Design.HintTooltips.titleText)
                         .padding(.bottom, 4)
                     
                     Text(desc)
                         .font(.custom(FontFamily.Inter.medium.name, size: 14))
-                        .zForegroundColor(Design.Text.primary)
+                        .zForegroundColor(Design.HintTooltips.bodyText)
                         .lineLimit(nil)
                         .lineSpacing(1.5)
                 }