+++ CHANGELOG.md
@@ -8,6 +8,7 @@ directly impact users rather than highlighting other crucial architectural updat
 
 ### Fixed
 - Crash in the background task during the overnight sync.
+- Tooltip font colors.
 
 ## 1.4 build 2 (2025-03-06)